Update README.md
steps 1-6  which includes telling the model to use the function_calls. That should not be necessary as it is very similar to the actual prompt we create. If you go to the Advanced Prompts section in the console you can see our prompts and your instructions are almost duplicates of what we put in.